+++ src/extract.rs
@@ -2,3 +2,4 @@
 
 pub use crate::json::{Json, DEFAULT_JSON_LIMIT};
 pub use crate::lazy_data::LazyData;
+pub use crate::query::Query;