+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 
 ## Unreleased - 2021-xx-xx
+- Copy `Query` extractor from Actix Web that can track minor versions of `serde-urlencoded`.
 
 
 ## 0.8.0 - 2022-01-20